+++ client/src/components/navbar/navbar.js
@@ -1,7 +1,7 @@
 import Container from 'react-bootstrap/Container';
 import Nav from 'react-bootstrap/Nav';
 import Navbar from 'react-bootstrap/Navbar';
-import { Link } from 'react-router-dom';
+import { NavLink } from 'react-bootstrap';
 import "./navbar.css"
 
 function Header () {
@@ -9,20 +9,20 @@ function Header () {
     <Navbar bg="light" expand="lg">
       <Container>
         <Navbar.Brand className="navbar-brand" href="#">
-          <Nav.Link><Link className="nav-link" to="/">
-            <span className="text-warning p-2">T|O</span>Dental</Link></Nav.Link>
+          <NavLink className="nav-link" href="/">
+            <span className="text-warning p-2">T|O</span>Dental</NavLink>
         </Navbar.Brand>
-        <Nav.Link><Link className="instagram nav-link p-3" to="https://www.instagram.com/todenta1/"><i class="bi bi-instagram"></i></Link></Nav.Link>
-        <Nav.Link><Link className="facebook nav-link" to="https://www.facebook.com/gkutaladze"><i class="bi bi-facebook"></i></Link></Nav.Link>
+        <NavLink className="instagram nav-link p-3" href="https://www.instagram.com/todenta1/" target="_blank"><i className="bi bi-instagram"></i></NavLink>
+        <NavLink className="facebook nav-link" href="https://www.facebook.com/gkutaladze" target="_blank"><i className="bi bi-facebook"></i></NavLink>
         <Navbar.Toggle aria-controls="basic-navbar-nav" />
         <Navbar.Collapse id="basic-navbar-nav">
           <Nav className="nav ms-auto">
-            <Nav.Link><Link className="nav-link" to="/">Home</Link></Nav.Link>
-            <Nav.Link><Link className="nav-link" to="/about">About</Link></Nav.Link>
-            <Nav.Link><Link className="nav-link" to="/equipments">Equipment</Link></Nav.Link>
-            <Nav.Link><Link className="nav-link" to="/newPatients">New Patients</Link></Nav.Link>
-            <Nav.Link><Link className="nav-link" to="/login">Login</Link></Nav.Link>
-            <Nav.Link><Link className="nav-link" to="/dashboard">Dashboard</Link></Nav.Link>
+            <NavLink className="nav-link" href="/">Home</NavLink>
+            <NavLink className="nav-link" href="/about">About</NavLink>
+            <NavLink className="nav-link" href="/equipments">Equipment</NavLink>
+            <NavLink className="nav-link" href="/newPatients">New Patients</NavLink>
+            <NavLink  className="nav-link" href="/login">Login</NavLink>
+            <NavLink  className="nav-link" href="/dashboard">Dashboard</NavLink>
           </Nav>
         </Navbar.Collapse>
       </Container>
