+++ server/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "nodemon server.js",
+    "start": "node server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
@@ -15,9 +15,7 @@
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "mysql2": "^3.2.3",
-    "nodemailer": "^6.9.1"
-  },
-  "devDependencies": {
+    "nodemailer": "^6.9.1",
     "nodemon": "^2.0.22"
   }
 }
