+++ server/server.js
@@ -3,8 +3,10 @@ const express = require("express");
 const app = express();
 const router = require("./routes/router")
 const cors = require("cors")
+const mysql2 = require("mysql2");
+require("./database");
 
-const port = 3001;
+const port = 4001;
 
 app.use(express.json());
 app.use(cors());
